商品削除機能を出品者のみに制限

+++ app/controllers/items_controller.rb
@@ -36,7 +36,7 @@ class ItemsController < ApplicationController
 
   def destroy
     item = Item.find(params[:id])
-    item.destroy
+    item.destroy if (current_user == item.user) && user_signed_in?
     redirect_to root_path
   end
